+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Mapbox welcomes participation and contributions from everyone.
 ## main
 
 ## 11.23.0-rc.1 - 20 April, 2026
+* Use TileStore::setRootPath(path) and TileStore::create() instead of deprecated TileStore::create(path).
+* Add setRootPath method to TileStore swift wrapper.
 
 ## 11.20.3 - 14 April, 2026
 ## 11.22.0 - 16 April, 2026